requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncclick==8.3.0.3
attrs==25.4.0
bcrypt==5.0.0
blinker==1.6.3
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
exceptiongroup==1.3.1
Flask==3.0.3
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
netifaces==0.11.0
priority==2.0.0
propcache==0.4.1
pycparser==2.23
pydantic==1.10.26
pydantic_core==2.41.5
python-kasa==0.7.7
Quart==0.19.9
quart-auth==0.11.0
quart-sqlalchemy==3.0.4
sniffio==1.3.1
SQLAlchemy==2.0.46
SQLAlchemy-Utils==0.42.1
sqlapagination==0.0.1
taskgroup==0.2.2
tenacity==8.0.1
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.0.6
wsproto==1.3.2
yarl==1.22.0
